refactor: alterando os arquivos models para pasta adm
