+++ src/models/admin/Category.ts
@@ -1,4 +1,4 @@
-import Database from './Database'
+import Database from './../Database'
 
 class Category extends Database {
   protected db = {
