+++ src/models/admin/User.ts
@@ -1,5 +1,5 @@
 /* eslint-disable camelcase */
-import Database from './Database'
+import Database from './../Database'
 
 class User extends Database {
   protected db = {
